requirements.txt
pandas==2.0.1
plotly==5.14.1
Pyrebase4==4.7.0
streamlit==1.22.0
streamlit_option_menu==0.3.4
urllib3==1.26.15
yfinance==0.2.18
